2단계 - 문자열 덧셈 계산기 코드리뷰 반영
